New module build and template
This now makes use of the tachyons-cli to do the heavy lifting for the
build and readme generation.

This implementation no longer has built in file watching for builds,
however everything is handled with the npm start command so it keeps
everything pretty simple.

I've also created a basic package.json template and format for the
modules to ensure that they can be consistent and efficient amongst all
modules. The style attribute now points to the source css so that it
plays nicer with postcss build systems and now specifies the files
attribute to ensure only the src and css directories are downloaded on
npm install.

Something that definitely needs to be fixed is the use of the
tachyons-cli dependency. Right now I'm specifying ../../package.json
because the execution context is from within the
node_modules/tachyons-cli directory, so we need to clean that up some
how. ¯\_(ツ)_/¯

:heart: